requirements.txt
beautifulsoup4==4.8.2
ioc-fanger==3.0.18
ioc-finder==2.0.13
OTXv2==1.5.6
pandas==1.0.3
pytorch-pretrained-bert==0.6.2
PyYAML==5.3
requests==2.22.0
scikit-learn==0.21.2
torch==1.3.0  # This version might be not compatible with your environment. Please install torch compatible with your environment.
Keras==2.2.4
tensorflow==1.13.2
numpy==1.18.1
spacy==2.2.3
xgboost==0.90
